Fix crash in resize
TODO: This probably is not an optimal solution. I have no idea why we
can't get a write lock to the uniform buffer there. The current "fix"
is to just set the dimensions every frame, because in the
`render_frame` function we appearntly are able to get a write lock to
the uniform buffer.